[iOS] Shift protection stats logic into a tab helper
This moves the logic into a location that can be used by the migrated script as well in a future PR and cleans up a few small things
